ElixirNeuralNetwork (Elixir Neurales Netzwerk v0.1.0)

Hauptmodul für die Kernlogik des neuralen Netzwerkes.

Functions

build(input_shape)

Erstellt ein neurales Netz.

Das Netz hat 28x28 Input-Neuronen, 128 "versteckte" Neuronen und 10 Ergebnis-Neuronen. Als Aktivierungs-Funktion wird hier für beide Ebenen die Logistische Funktion $\sigma(x)$ verwendet.

display_data(arg)

display_image(image)

display_network(model)

download()

Importiert den MNIST-Datensatz

Der MNIST-Datensatz besteht aus 60.000 28x28 Pixel großen Bildern von handgeschriebenen Zahlen. (https://en.wikipedia.org/wiki/MNIST_database)

Die Daten werden in einem "Tuple", ein Container für eine bestimmte Anzahl von Elementen, gespeichert.

Format

iex> download()
{images, labels}

load!(path)

Lädt das Netzwerk und dessen Zusaznd aus der angegebenen Datei.

predict(arg, data)

Stellt das Netzwerk als Tebelle dar.

prediction_images(arg)

save!(data, path)

Speichert das Netzwerk und dessen Zusaznd in der angegebenen Datei.

test(model, model_state, test_images, test_labels)

train(model, train_images, train_labels, epochs)

Trainiert das neurale Netz.

Hierzu wird es über mehrere Generationen "Epochen" mit dem Trainingsdatensatz trainiert.

transform_images(arg, split)

transform_labels(arg, split)