ElixirNeuralNetwork.Network (Elixir Neurales Netzwerk v0.1.0)

Modul für die Kernlogik des neuralen Netzwerkes.

Functions

download()

Importiert den MNIST-Datensatz

iex> download()
{images, labels}

gradient_descent(network, training_data, epochs, learning_rate)

init(sizes)

@spec init(list()) :: %ElixirNeuralNetwork.Network{
  biases: list(),
  num_layers: non_neg_integer(),
  sizes: list(),
  weights: list()
}

Initiert die Werte des neuralen Netzwerks.

Als Parameter nimmt die Funktion die Dimensionen, die das Netz annehmen soll.

Beispiel

iex> init_network([2, 4, 2])
%Network{num_layers: 3, sizes: [2, 4, 2], bias: [...], weights: [...]}

load!()

new()

path()

prepare_data(batch_size)

@spec prepare_data(pos_integer()) ::
  {[{Nx.Tensor, Nx.Tensor}], [{Nx.Tensor, Nx.Tensor}], [{Nx.Tensor, Nx.Tensor}]}

Bereitet den Datensatz zur Verarbeitung im neuralen Netz vor.

Teilt die Daten in Trainings- Validations- und Testdaten ein.

iex> transform_images({images, labels})
{training_data, validation_data, test_data}

save!(model, state)

test(model, test_data)

train(model, training_data, validation_data)

transform_images(arg)

transform_labels(arg)